requirements.txt
opencv_python==4.0.0.21
numpy==1.15.2
pandas==0.23.4
torch==0.4.1
Pillow==6.0.0